+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Fixed: For local deploy, start with 0.0.0 if module does not exist
+
 ## 2.0.0 - 2019-11-07
 
 * Added: Invoke-BuildIsolated command for isolated builds